# projet-08-Ot-gether-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ot'gether API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/events/{event}/participate": {
            "post": {
                "tags": [
                    "Participate"
                ],
                "summary": "Add User Auth paticipate at an Event",
                "description": "Add User Auth paticipate at an Event.",
                "operationId": "29217c904c6350f358d471c5c461f984",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of event to participate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User participate at an event"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Participate"
                ],
                "summary": "Remove User Auth paticipate at an Event",
                "description": "Remove User Auth paticipate at an Event.",
                "operationId": "c35d70f1269aad7c28b618e30a080711",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of event to participate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User participate at an event"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
        "/api/events/{event}/report": {
            "post": {
                "tags": [
                    "Report"
                ],
                "summary": "Report an event",
                "operationId": "32f0ab6c76bc9739ce1eab37566af720",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of event to report",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "Message to send with the report",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event reported"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
        "/api/tags": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Get list of tags",
                "operationId": "65f85cc85cd13d2454b6c88ffff5cc3a",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tag"
                ],
                "summary": "Register a new tag",
                "operationId": "467bf3e5824fd78f9cdfb70c0ba9b172",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tag's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tag created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/tags/{tag}": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Get a tag by id",
                "operationId": "890706647a0186ae7289ff232e056564",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "description": "ID of tag to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tag"
                ],
                "summary": "Delete a tag by id",
                "operationId": "d3741650499e47b0c4644657896e80bc",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "description": "ID of tag to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tag"
                ],
                "summary": "Update a tag by id",
                "operationId": "bbe383346f5285711681b6e399a1137e",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "description": "ID of tag to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tag name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}